requirements.txt
Django==2.1.1
djangorestframework==3.8.2
django-cors-headers==2.4.0
